+++ render/tests/index.html
@@ -21,6 +21,7 @@
 		<script src="test-fragment.js"></script>
 		<script src="test-normalize.js"></script>
 		<script src="test-normalizeChildren.js"></script>
+		<script src="test-normalizeComponentChildren.js"></script>
 		<script src="test-createText.js"></script>
 		<script src="test-createHTML.js"></script>
 		<script src="test-createFragment.js"></script>
